+++ lib/src/extended_network_image_utils.dart
@@ -1,3 +1,4 @@
+import 'dart:async';
 import 'dart:convert';
 import 'dart:io';
 import 'dart:typed_data';
@@ -8,7 +9,7 @@ import 'package:path_provider/path_provider.dart';
 
 import 'extended_network_image_provider.dart';
 
-const String CacheImageFolderName = "cacheimage";
+const String cacheImageFolderName = "cacheimage";
 
 String keyToMd5(String key) => md5.convert(utf8.encode(key)).toString();
 
@@ -16,14 +17,14 @@ String keyToMd5(String key) => md5.convert(utf8.encode(key)).toString();
 ///  <param name="duration">timespan to compute whether file has expired or not</param>
 Future<bool> clearDiskCachedImages({Duration duration}) async {
   try {
-    Directory _cacheImagesDirectory = Directory(
-        join((await getTemporaryDirectory()).path, CacheImageFolderName));
-    if (_cacheImagesDirectory.existsSync()) {
+    final cacheImagesDirectory = Directory(
+        join((await getTemporaryDirectory()).path, cacheImageFolderName));
+    if (cacheImagesDirectory.existsSync()) {
       if (duration == null) {
-        _cacheImagesDirectory.deleteSync(recursive: true);
+        cacheImagesDirectory.deleteSync(recursive: true);
       } else {
         var now = DateTime.now();
-        for (var file in _cacheImagesDirectory.listSync()) {
+        for (var file in cacheImagesDirectory.listSync()) {
           FileStat fs = file.statSync();
           if (now.subtract(duration).isAfter(fs.changed)) {
             //print("remove expired cached image");
@@ -44,7 +45,7 @@ Future<bool> clearDiskCachedImage(String url) async {
   try {
     var file = await getCachedImageFile(url);
     if (file != null) {
-      file.delete(recursive: true);
+      await file.delete(recursive: true);
     }
   } catch (_) {
     return false;
@@ -57,10 +58,10 @@ Future<bool> clearDiskCachedImage(String url) async {
 Future<File> getCachedImageFile(String url) async {
   try {
     var key = keyToMd5(url);
-    Directory _cacheImagesDirectory = Directory(
-        join((await getTemporaryDirectory()).path, CacheImageFolderName));
-    if (_cacheImagesDirectory.existsSync()) {
-      for (var file in _cacheImagesDirectory.listSync()) {
+    final cacheImagesDirectory = Directory(
+        join((await getTemporaryDirectory()).path, cacheImageFolderName));
+    if (cacheImagesDirectory.existsSync()) {
+      for (var file in cacheImagesDirectory.listSync()) {
         if (file.path.endsWith(key)) {
           return file;
         }
@@ -82,55 +83,26 @@ ImageCache getMemoryImageCache() {
   return PaintingBinding.instance.imageCache;
 }
 
-//List<ExtendedNetworkImageProvider> pendingImages =
-//    List<ExtendedNetworkImageProvider>();
+/// get network image data from cached
+Future<Uint8List> getNetworkImageData(
+  String url, {
+  bool useCache = true,
+  StreamController<ImageChunkEvent> chunkEvents,
+}) async {
+  return ExtendedNetworkImageProvider(url, cache: useCache).getNetworkImageData(
+    chunkEvents: chunkEvents,
+  );
+}
 
-//void cancelPendingNetworkImageByToken(CancellationToken cancelToken,
-//    {bool takeCareSameUrl: true}) {
-//  if (cancelToken == null) return;
-//
-//  if (!takeCareSameUrl) cancelToken.cancel();
-//
-//  pendingImages
-//      .where((image) => image.cancelToken == cancelToken)
-//      ?.forEach((f) {
-//    cancelPendingNetworkImageByProvider(f, takeCareSameUrl: takeCareSameUrl);
-//  });
-//}
-//
-//void cancelPendingNetworkImageByProvider(ExtendedNetworkImageProvider provider,
-//    {bool takeCareSameUrl: true}) {
-//  if (provider == null) return;
-//
-//  if (!takeCareSameUrl) provider.cancelToken?.cancel();
-//
-//  ///find the same image(url,scale) with different cancel token
-//  var pendingImageList = pendingImages.where((image) =>
-//      image == provider && image.cancelToken != provider.cancelToken);
-//
-//  if (pendingImageList.length > 0) {
-//    pendingImageList.forEach((image) {
-//      ///may it has no cancel token, or the request is not canceled
-//      bool shouldCancel = image.cancelToken?.isCanceled == true;
-//      print(shouldCancel);
-//
-//      ///if any one is not cancel break, this image should not be canceled.
-//      if (!shouldCancel) return;
-//    });
-//  }
-//  provider.cancelToken?.cancel();
-//}
-
-//void cancelPendingNetworkImageByUrl(String url, {bool takeCareSameUrl: true}) {
-//  if (url == null) false;
-//
-//  pendingImages.where((image) => image.url == url).forEach((f) {
-//    cancelPendingNetworkImageByProvider(f, takeCareSameUrl: takeCareSameUrl);
-//  });
-//}
-
-///get network image data from cached
-Future<Uint8List> getNetworkImageData(String url, {bool useCache: true}) async {
-  return ExtendedNetworkImageProvider(url)
-      .getNetworkImageData(useCache: useCache);
+/// get total size of cached image
+Future<int> getCachedSizeBytes() async {
+  int size = 0;
+  final cacheImagesDirectory = Directory(
+      join((await getTemporaryDirectory()).path, cacheImageFolderName));
+  if (cacheImagesDirectory.existsSync()) {
+    for (var file in cacheImagesDirectory.listSync()) {
+      size += file.statSync().size;
+    }
+  }
+  return size;
 }