+++ CHANGELOG.md
@@ -1,7 +1,13 @@
+## [0.2.2]
+
+* add analysis_options.yaml
+* fix null exception of chunkEvents
+
 ## [0.2.1]
 
 * support loading progress for network
 * public HttpClient of ExtendedNetworkImageProvider
+* add getCachedSizeBytes method
   
 ## [0.2.0]